jump search

+++ 0x1E-search_algorithms/100-jump.c
@@ -18,11 +18,10 @@ int jump_search(int *array, size_t size, int value)
 	if (array == NULL || size == 0)
 		return (-1);
 
-	step = (int)sqrt((double)size);
+	step = sqrt(size);
 	current = prev = 0;
 
-	do
-	{
+	do {
 		printf("Value checked array[%d] = [%d]\n", current, array[current]);
 
 		if (array[current] == value)
